Fix game bug where buckets are consumed in recipes
Buckets are not consumed when milk is drank from a bucket and shouldn't be consumed when cooking either. This is broken in the base game, but can be easily fixed in this mod by adding an empty bucket to the players inventory when they cook a recipe that used a bucket of milk.